refactor: extract ownership logic to new file
